Use namespaces in renderers
